backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
supabase==1.0.4
requests==2.31.0
python-dateutil==2.8.2
pandas==2.0.3
openpyxl==3.1.2
xlrd==2.0.1
